allow custom match report class to be passed to find_matches command
